First refactors regarding trajectories router
