+++ src/app/trajectories/[id]/components/Sidebar.tsx
