--- src/app/trajectories/page.tsx
@@ -1,3 +1,0 @@
-export default async function TrajectoryPage() {
-  return <main />;
-}
